chore: move errors to custom file
